+++ frontend/src/api/v3/vuetify-items.js
@@ -1,11 +1,6 @@
 // Shared functions for most metadata components.
-import { vuetifyNullCode } from "@/choices/browser-choices.json";
-export const NULL_PKS = new Set([
-  "",
-  vuetifyNullCode,
-  vuetifyNullCode.toString(),
-  undefined,
-]);
+import { vuetifyNullCode as VUETIFY_NULL_CODE } from "@/choices/browser-choices.json";
+export const NULL_PKS = new Set(["", VUETIFY_NULL_CODE, undefined, null]);
 
 const toVuetifyItem = function (item) {
   /*
@@ -25,7 +20,7 @@ const toVuetifyItem = function (item) {
         vuetifyItem = { value, title: item.name };
       }
     } else if (NULL_PKS.has(item.pk)) {
-      vuetifyItem = undefined;
+      vuetifyItem = { value: item.pk, title: "None" };
     } else {
       vuetifyItem = { value: item.pk, title: item.name };
     }
@@ -60,6 +55,7 @@ export const toVuetifyItems = function ({
 
   // Case insensitive search for filter-sub-menu
   const lowerCaseFilter = filter ? filter.toLowerCase() : filter;
+  let noneItem;
 
   let computedItems = [];
   for (const item of sourceItems) {
@@ -69,12 +65,21 @@ export const toVuetifyItems = function ({
       (!lowerCaseFilter ||
         vuetifyItem?.title?.toLowerCase().includes(lowerCaseFilter))
     ) {
-      computedItems.push(vuetifyItem);
+      if (NULL_PKS.has(vuetifyItem.value)) {
+        noneItem = vuetifyItem;
+        noneItem.value = VUETIFY_NULL_CODE;
+      } else {
+        computedItems.push(vuetifyItem);
+      }
     }
   }
   if (sort) {
     const sortFunc = numeric ? vuetifyItemCompareNumeric : vuetifyItemCompare;
     computedItems = computedItems.sort(sortFunc);
+  }
+  if (noneItem) {
+    // Prepend noneItem if it exists.
+    computedItems.unshift(noneItem);
   }
   return computedItems;
 };